moved graphviz_present to conftest and check in cmake test
