+++ tests/test_cmake.py
@@ -17,6 +17,7 @@ from contextlib import contextmanager
 import pytest
 import toml
 
+from conftest import graphviz_present
 from loki import (
     gettempdir, execute
 )
@@ -204,6 +205,8 @@ def test_cmake_plan(srcdir, config, cmake_project, loki_install, ecbuild, silent
 
             # Make sure the plan files have been created
             assert (builddir/'loki_plan.cmake').exists()
+            if graphviz_present():
+                assert (builddir/'loki_callgraph.pdf').exists()
 
             # Validate the content of the plan file
             loki_plan = (builddir/'loki_plan.cmake').read_text()
